feat: modularized STATUS
